+++ Sources/RichTextKit/Extensions/NSAttributedString+Empty.swift
@@ -3,7 +3,7 @@
 //  RichTextKit
 //
 //  Created by Daniel Saidi on 2022-06-01.
-//  Copyright © 2022-2023 Daniel Saidi. All rights reserved.
+//  Copyright © 2022-2024 Daniel Saidi. All rights reserved.
 //
 
 import Foundation
@@ -12,6 +12,6 @@ public extension NSAttributedString {
 
     /// Create an empty attributed string.
     static var empty: NSAttributedString {
-        NSAttributedString(string: "")
+        .init(string: "")
     }
 }