+++ Sources/RichTextKit/Extensions/String+Characters.swift
@@ -3,37 +3,42 @@
 //  RichTextKit
 //
 //  Created by Daniel Saidi on 2022-05-29.
-//  Copyright © 2022-2023 Daniel Saidi. All rights reserved.
+//  Copyright © 2022-2024 Daniel Saidi. All rights reserved.
 //
 
 import Foundation
 
 public extension String.Element {
 
-    /// Get the string element for a `\r` carriage return.
+    /// An `\r` carriage return string element.
     static var carriageReturn: String.Element { "\r" }
 
-    /// Get the string element for a `\n` newline.
+    /// An `\n` newline string element.
     static var newLine: String.Element { "\n" }
 
-    /// Get the string element for a `\t` tab.
+    /// A `\t` tab string element.
     static var tab: String.Element { "\t" }
 
-    /// Get the string element for a ` ` space.
+    /// A ` ` space string element.
     static var space: String.Element { " " }
+    
+    /// If the string element is a new line separator.
+    var isNewLineSeparator: Bool {
+        self == .newLine || self == .carriageReturn
+    }
 }
 
 public extension String {
 
-    /// Get the string for a `\r` carriage return.
+    /// An `\r` carriage return string.
     static let carriageReturn = String(.carriageReturn)
 
-    /// Get the string for a `\n` newline.
+    /// An `\n` newline string.
     static let newLine = String(.newLine)
 
-    /// Get the string for a `\t` tab.
+    /// A `\t` tab string.
     static let tab = String(.tab)
 
-    /// Get the string for a ` ` space.
+    /// A ` ` space string.
     static let space = String(.space)
 }