+++ Sources/RichTextKit/Extensions/String+Paragraph.swift
@@ -3,7 +3,7 @@
 //  RichTextKit
 //
 //  Created by Daniel Saidi on 2022-05-29.
-//  Copyright © 2022-2023 Daniel Saidi. All rights reserved.
+//  Copyright © 2022-2024 Daniel Saidi. All rights reserved.
 //
 
 import Foundation
@@ -11,11 +11,8 @@ import Foundation
 public extension String {
 
     /**
-     Backs to find the index of the first new line paragraph
-     before the provided location, if any.
-
-     A new paragraph is considered to start at the character
-     after the newline char, not the newline itself.
+     Look backward to find the index of the paragraph before
+     the provided location, if any.
      */
     func findIndexOfCurrentParagraph(from location: UInt) -> UInt {
         if isEmpty { return 0 }
@@ -31,12 +28,8 @@ public extension String {
     }
 
     /**
-     Looks forward to find the next new line paragraph after
-     the provided location, if any. If no next paragraph can
-     be found, the current is returned.
-
-     A new paragraph is considered to start at the character
-     after the newline char, not the newline itself.
+     Look forward to find the index of a paragraph after the
+     provided location, if any.
      */
     func findIndexOfNextParagraph(from location: UInt) -> UInt {
         var index = location
@@ -51,14 +44,9 @@ public extension String {
     }
 
     /**
-     Looks forward to find the next new line paragraph after
-     the provided location, if any. If no next paragraph can
-     be found, the last index of the paragraph is returned.
-
-     A new paragraph is considered to start at the character
-     after the newline char, not the newline itself.
+     Look forward to find the index of a paragraph after the
+     provided location, if any.
      */
-
     func findIndexOfNextParagraphOrEndOfCurrent(from location: UInt) -> UInt {
         var index = location
         repeat {
@@ -72,10 +60,7 @@ public extension String {
     }
 
     /**
-     Returns the length of the paragraph found at the location provided..
-
-     A new paragraph is considered to start at the character
-     after the newline char, not the newline itself.
+     Get the length of the paragraph at a provided location.
      */
     func findLengthOfCurrentParagraph(from location: UInt) -> Int {
         if isEmpty { return 0 }
@@ -85,7 +70,7 @@ public extension String {
     }
 
     /**
-     Returns the index of the word at the location provided..
+     Get the index of the word at the provided text location.
 
      A word is considered to be a length of text between two
      breaking characters or space characters.